requirements.txt
albumentations==1.4.21
black==24.10.0
fastapi==0.115.6
httpx==0.28.1
imagehash==4.3.1
joblib==1.4.2
lime==0.2.0.1
matplotlib==3.9.3
numpy==1.26.4
opencv-python-headless==4.10.0.84
pandas==2.2.3
pillow==11.0.0
pre-commit==4.0.1
pydantic==2.10.4
pydantic-settings==2.7.0
pytest==8.3.4
python-multipart==0.0.20
pytorch-grad-cam==1.5.4
PyYAML==6.0.2
reportlab==4.2.5
ruff==0.8.4
scikit-image==0.24.0
scikit-learn==1.6.0
scipy==1.14.1
shap==0.46.0
streamlit==1.41.1
timm==1.0.12
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
uvicorn[standard]==0.34.0
wandb==0.19.1
